Expose credit card verifications when present
When a customer is created with the verify_card flag, the results of
credit card verification are included in the XML response, nested within
the credit card.

This change exposes the verification data, which was previously being
discarded.

The verification data is useful for apps that may wish to, for example,
store the AVS and CVV results for fraud analysis and chargeback response
without having to separately look up those values from the Braintree
dashboard.