sharding vVOlRllG - Confirm new data on server does not impact pagination
Transaction and Verification searches will be paginated across shards
in a multisharded world.  This increases the scenarios where results
may be duplicated/skipped during paging.

This does not appear to be a problem in Client Libraries, because of the
two-step process to retrieve results. Created basic specs to confirm this
is the case.