Adjust APW specs to work with stubbed responses
Previously when this spec was run, we would actually make calls to
Apple's API since we had no stubbing in the Gateway. This led to
intermittent failures when Apple's environment had isssues.

We are working on a change to make it so in the test environment, we
stub all calls to Apple
(https://github.braintreeps.com/braintree/gateway/pull/6948). Once this
is merged, the behavior here will be slightly changed since we will be
returning a stateless constant stubbed response for each test.

This commit tests that we make the requests correctly and that we
receive the expected stubbed response from the Gateway. We will also
apply this change retroactively to every build between when this spec
was originally added and now so that the CI builds continue to pass.